DDP-8589 refactoring
