+++ pepper-apis/dss-core/src/main/java/org/broadinstitute/ddp/model/copy/UserType.java
@@ -1,0 +1,6 @@
+package org.broadinstitute.ddp.model.copy;
+
+public enum UserType {
+    USER,
+    OPERATOR;
+}
