+++ pepper-apis/studybuilder-cli/src/main/java/org/broadinstitute/ddp/studybuilder/EventBuilder.java
@@ -40,6 +40,7 @@ import org.broadinstitute.ddp.model.copy.CopyConfigurationPair;
 import org.broadinstitute.ddp.model.copy.CopyLocation;
 import org.broadinstitute.ddp.model.copy.CopyLocationType;
 import org.broadinstitute.ddp.model.copy.CopyPreviousInstanceFilter;
+import org.broadinstitute.ddp.model.copy.UserType;
 import org.broadinstitute.ddp.model.dsm.DsmNotificationEventType;
 import org.broadinstitute.ddp.model.event.ActivityStatusChangeTrigger;
 import org.broadinstitute.ddp.model.event.DsmNotificationTrigger;
@@ -51,7 +52,6 @@ import org.broadinstitute.ddp.model.user.EnrollmentStatusType;
 import org.broadinstitute.ddp.model.workflow.ActivityState;
 import org.broadinstitute.ddp.model.workflow.StateType;
 import org.broadinstitute.ddp.model.workflow.StaticState;
-import org.broadinstitute.ddp.pex.UserType;
 import org.broadinstitute.ddp.util.ConfigUtil;
 import org.jdbi.v3.core.Handle;
 
@@ -361,7 +361,8 @@ public class EventBuilder {
     private CopyLocation buildCopyLocation(Config locationCfg) {
         var type = CopyLocationType.valueOf(locationCfg.getString("type"));
         if (type == CopyLocationType.ANSWER) {
-            String user = locationCfg.hasPath("user") ? locationCfg.getString("user") : UserType.USER.toUpperCase();
+            UserType user = locationCfg.hasPath("user") ? UserType.valueOf(locationCfg.getString("user"))
+                    : UserType.USER;
             return new CopyAnswerLocation(locationCfg.getString("questionStableId"), user);
         } else {
             return new CopyLocation(type);
